Don't call an event function twice
In case one event function was added for read and write events of the same
source it could be called twice, once for the read event and once for the
write event.

Also recheck the event source status after calling the read event function
and before the write event function in case the read event function removed
the event source.