Remove all log categories
They never worked as intended. For normal use one would just set all
categories to info or warn and be done with it. In this case there is
no need to filter the log by category as it should not be that large
anyway. The main purpose of the log under normal usage conditions is
to see if any errors or warnings occurred at all. For this use case
one should just look at all messages equally, no categories necessary.

For debugging one would set all categories to debug. This produces
large amounts of messages. Depending on which part of the daemon should
be debugged it would be helpful to be able to filter the debug messages
but the category system is way to coarse for this and not easy enough
to change, as it can only be controlled via <daemon>.conf. So categories
don't help here either.

That's why the categories get completely removed now. They are either
not necessary or not good enough. For debugging a more fine grained and
easier to use system will be added later.