Add debug log filters
Most of the log messages are on debug level. The new filter system
allows to filter (include/exclude) debug messages based on their
source. The source can be the name of the file or a group name:

  all, common, event, packet, object

A debug filter can be specified in the config file or via the --debug
command line option. For example:

  brickd --debug -event,-packet,+writer.c

excludes all debug message from the event and packet group, but includes
all debug message from the writer.c file, even if they are in the event
or packet group.