Fix packet duplication in writers partial write logic
If the packet was partially written then the remaining packet was
pushed to the backlog. But after the push operation a return 1 statement
was missing so complete packet was pushed to the backlog a second time.
In the end this resulted in sending the packet twice.

Simplify and fix the partial write logic.