writer: Avoid accessing freed memory
The log_debug call accesses the partial_packet that was peeked from
the queue. The queue owns the memory of partial_packet and frees it
in the queue_pop call. Therefore, queue_pop cannot be called before
the log_debug call, as this would make the log_debug access freed
memory.