Decouple log_<level> calls from actual output writing
Use a 256kB ringbuffer and a thread for the decoupling. This is done
to avoid blocking the log_<level> call any longer then necessary.
Outputting log messages is especially slow when running on Windows
in cmd.exe. It is so slow that is messes with libusb event handling
timing if debug level logging is enabled. This is avoided by the
decoupling because the only thing left to be done under the global
logging lock is adding the log message to the ringbuffer.